third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/reporting/access-reporting/reporting-observer-expected.txt

This is a testharness.js-based test.
FAIL Opener COOP assert_equals: No report received. expected 1 but got 0
FAIL Openee COOP assert_equals: No report received. expected 1 but got 0
FAIL Access from same-origin iframe assert_not_equals: No report received. got disallowed value "timeout"
PASS Access from cross-site iframe
Harness: the test ran to completion.
